requirements.txt
dynaconf>=3.1.2
starlette==0.14.2
uvicorn[standard]>=0.13.4
gunicorn>=20.0.4
itsdangerous>=1.1.0
jinja2>=2.11.0
passlib>=1.7.4
toml>=0.10.2
SQLAlchemy>=1.4.23
authlib>=1.0.0b2
dnspython>=2.1.0
python-multipart>=0.0.5
asgi-ratelimit>=0.4.0
aredis>=1.1.8
httpx>=0.21.1
cryptography>=36.0.0
aiohttp>=3.8.1
psutil>=5.8.0
